+++ docs/readme.org
@@ -28,8 +28,10 @@ CLOSED: [2019-06-19 Wed 11:05]
 *** [ ] Refactoring
 **** DONE Fix proxying host & port
 CLOSED: [2019-06-19 Wed 19:40]
-**** [ ] Migrating script for testing
-**** [ ] Make Backend-Service
+**** DONE Migrating script for testing
+CLOSED: [2019-06-20 Thu 15:12]
+**** DONE Make Backend-Service & Test it
+CLOSED: [2019-06-20 Thu 15:12]
 **** [ ] Fix responsiveness (Container & Grid & Box issues)
 **** [ ] Change rate from 1000 to 10,000
 *** [ ] Go back to login page when change wallet
